Dry Run mode

+++ clos/docker_startup_clos.sh
@@ -23,6 +23,7 @@ spine_groups=2
 leaf=8
 fl_containers=0
 hosts=0
+dry_run=0
 declare -A a_cores
 declare -A a_spines
 declare -A a_leaves
@@ -47,6 +48,7 @@ function display_help {
   echo "${REV}-t${NORM} Specifies the number of ${BOLD}leaf${NORM}. Default is ${BOLD}$leaf${NORM}."
   echo "${REV}-f${NORM} Specifies the filename in which the created containers will be recorded. Default is ${BOLD}$container_record${NORM}."
   echo "${REV}-z${NORM} Specifies the number of ${BOLD}hosts${NORM}. Default is ${BOLD}$hosts${NORM}."
+  echo "${REV}-d${NORM} Do a ${BOLD}dry run${NORM}. Print what would be done, but not actually do anything"
   echo -e "${REV}-h${NORM} Displays this help message. No further functions are performed."\\n
   echo -e "Example: ${BOLD}$self -s 2 -l 4 -t 8 -g 2 -z 0 ${NORM}"\\n
   exit 0
@@ -57,7 +59,7 @@ function display_help {
 # TODO: Actually accept these parameters on the command line...
 # For now, these are ignored, because I've not figured out a good algorithm to do the spine-group/leaf-group bits.
 # A better coder than I can do this, I suspect.
-optspec=":s:l:t:z:g:f:h"
+optspec=":s:l:t:z:g:f:hd"
 while getopts "$optspec" optchar; do
   case $optchar in
     s)
@@ -83,6 +85,10 @@ while getopts "$optspec" optchar; do
     f)
       container_record=$OPTARG
       echo "Container record ${REV}filename${NORM} is set to: ${BOLD}$container_record${NORM}" >&2
+      ;;
+    d)
+      dry_run=1
+      echo "${BOLD}Dry Run ${NORM} mode. ${BOLD}No changes will be made${NORM}" >&2
       ;;
     h)
       display_help
@@ -110,7 +116,9 @@ if [ -e "$container_record" ]; then
   echo "Cowardly refusing to continue"
   exit 1
 else
-  touch $container_record
+  if [ $dry_run -eq 0 ]; then 
+    touch $container_record
+  fi # if [ $dry_run -eq 1 ]
   if [ $? -ne 0 ]; then
     echo "${RED}ERROR: ${NORM} Container record file ${container_record} does not appear writeable"
     exit 1
@@ -118,11 +126,15 @@ else
 fi # if [ -e "$container_record" ]
 
 echo "***** Checkout the flexswitch base image *******"
-sudo docker pull snapos/flex:latest
-if [ $? -ne 0 ]; then
-	echo "${RED}ERROR:${NORM} Docker pull failed. Please check output above and fix" 1>&2
-	exit 1
-fi # if [ $? -ne 0 ]
+if [ $dry_run -eq 1 ]; then 
+  echo "${REV}DRY-RUN: ${NORM}sudo docker pull snapos/flex:latest"
+else 
+  sudo docker pull snapos/flex:latest
+  if [ $? -ne 0 ]; then
+	  echo "${RED}ERROR:${NORM} Docker pull failed. Please check output above and fix" 1>&2
+	  exit 1
+  fi # if [ $? -ne 0 ]
+fi
 
 echo -e "\n\n"
 
@@ -142,14 +154,19 @@ function docker_start_core {
   if [ ! -z "$1" ]; then
     instance="$1"
   fi # if [ ! -z "$1" ]; then
-  echo -e "\tStarting a docker container type: ${REV}FlexSwitch Core${NORM}, instance name: ${BOLD}core$instance${NORM}"
-  container_id=$(sudo docker run -dt --log-driver=syslog --privileged --cap-add ALL --hostname=core$instance --name core$instance -P snapos/flex:latest)
-  if [ $? -ne 0 ]; then
-    echo "${RED}ERROR:${NORM} Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
-    exit 1
+  if [ $dry_run -eq 1 ]; then
+    echo "${REV}DRY-RUN: ${NORM}"
+    echo -e "\tStarting a docker container type: ${REV}FlexSwitch Core${NORM}, instance name: ${BOLD}core$instance${NORM}"
   else
-    echo -n "$container_id," >> $container_record
-  fi # if [ $? -ne 0 ]
+    echo -e "\tStarting a docker container type: ${REV}FlexSwitch Core${NORM}, instance name: ${BOLD}core$instance${NORM}"
+  container_id=$(sudo docker run -dt --log-driver=syslog --privileged --cap-add ALL --hostname=core$instance --name core$instance -P snapos/flex:latest)
+    if [ $? -ne 0 ]; then
+      echo "${RED}ERROR:${NORM} Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
+      exit 1
+    else
+      echo -n "$container_id," >> $container_record
+    fi # if [ $? -ne 0 ]
+  fi # if [ $dry_run -eq 1 ]
   return 0
 }
 
@@ -157,14 +174,19 @@ function docker_start_spine {
   if [ ! -z "$1" ]; then
     instance="$1"
   fi # if [ ! -z "$1" ]; then
-  echo -e "\tStarting a docker container type: ${REV}FlexSwitch Spine${NORM}, instance name: ${BOLD}spine$instance${NORM}"
-  container_id=$(sudo docker run -dt --log-driver=syslog --privileged --cap-add ALL --hostname=spine$instance --name spine$instance -P snapos/flex:latest)
-  if [ $? -ne 0 ]; then
-    echo "${RED}ERROR:${NORM} Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
-    exit 1
+  if [ $dry_run -eq 1 ]; then
+    echo "${REV}DRY-RUN: ${NORM}"
+    echo -e "\tStarting a docker container type: ${REV}FlexSwitch Spine${NORM}, instance name: ${BOLD}spine$instance${NORM}"
   else
-    echo -n "$container_id," >> $container_record
-  fi # if [ $? -ne 0 ]
+    echo -e "\tStarting a docker container type: ${REV}FlexSwitch Spine${NORM}, instance name: ${BOLD}spine$instance${NORM}"
+    container_id=$(sudo docker run -dt --log-driver=syslog --privileged --cap-add ALL --hostname=spine$instance --name spine$instance -P snapos/flex:latest)
+    if [ $? -ne 0 ]; then
+      echo "${RED}ERROR:${NORM} Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
+      exit 1
+    else
+      echo -n "$container_id," >> $container_record
+    fi # if [ $? -ne 0 ]
+  fi # if [ $dry_run -eq 1 ]
   return 0
 }
 
@@ -172,14 +194,19 @@ function docker_start_leaf {
   if [ ! -z "$1" ]; then
     instance="$1"
   fi # if [ ! -z "$1" ]; then
-  echo -e "\tStarting a docker container type: ${REV}FlexSwitch Leaf${NORM}, instance name: ${BOLD}leaf$instance${NORM}"
-  container_id=$(sudo docker run -dt --log-driver=syslog --privileged --cap-add ALL --hostname=leaf$instance --name leaf$instance -P snapos/flex:latest)
-  if [ $? -ne 0 ]; then
-    echo "${RED}ERROR:${NORM} Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
-    exit 1
+  if [ $dry_run -eq 1 ]; then
+    echo "${REV}DRY-RUN: ${NORM}"
+    echo -e "\tStarting a docker container type: ${REV}FlexSwitch Leaf${NORM}, instance name: ${BOLD}leaf$instance${NORM}"
   else
-    echo -n "$container_id," >> $container_record
-  fi # if [ $? -ne 0 ]
+    echo -e "\tStarting a docker container type: ${REV}FlexSwitch Leaf${NORM}, instance name: ${BOLD}leaf$instance${NORM}"
+    container_id=$(sudo docker run -dt --log-driver=syslog --privileged --cap-add ALL --hostname=leaf$instance --name leaf$instance -P snapos/flex:latest)
+    if [ $? -ne 0 ]; then
+      echo "${RED}ERROR:${NORM} Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
+      exit 1
+    else
+      echo -n "$container_id," >> $container_record
+    fi # if [ $? -ne 0 ]
+  fi # if [ $dry_run -eq 1 ]
   return 0
 }
 
@@ -187,14 +214,19 @@ function docker_start_host {
   if [ ! -z "$1" ]; then
     instance="$1"
   fi # if [ ! -z "$1" ]; then
-  echo -e "\tStarting a docker container type: ${REV}Host${NORM}, instance name: ${BOLD}host$instance${NORM}"
-  container_id=$(sudo docker run -dt --log-driver=syslog --hostname=host$instance --name host$instance -P snapos/flex:latest)
-  if [ $? -ne 0 ]; then
-    echo "${RED}ERROR${NORM}Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
-    exit 1
+  if [ $dry_run -eq 1 ]; then
+    echo "${REV}DRY-RUN: ${NORM}"
+    echo -e "\tStarting a docker container type: ${REV}Host${NORM}, instance name: ${BOLD}host$instance${NORM}"
   else
-    echo -n "$container_id," >> $container_record
-  fi # if [ $? -ne 0 ]
+    echo -e "\tStarting a docker container type: ${REV}Host${NORM}, instance name: ${BOLD}host$instance${NORM}"
+    container_id=$(sudo docker run -dt --log-driver=syslog --hostname=host$instance --name host$instance -P snapos/flex:latest)
+    if [ $? -ne 0 ]; then
+      echo "${RED}ERROR${NORM}Failed starting docker instance \"$instance\". Please check output above and fix" 1>&2
+      exit 1
+    else
+      echo -n "$container_id," >> $container_record
+    fi # if [ $? -ne 0 ]
+  fi # if [ $dry_run -eq 1 ]
   return 0
 }
 
@@ -203,11 +235,15 @@ if [ $cores -ne 0 ]; then
   do
     echo "Core number: $i"
     docker_start_core $i
-    namespace=$(sudo docker inspect -f '{{.State.Pid}}' core$i)
-    mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' core$i)
-    echo "core$i,$namespace,$mgmt_ip" >> $container_record
-    echo -en "\t"; sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
-    a_cores["core$i"]="$namespace"
+    if [ $dry_run -eq 1 ]; then
+      echo "${REV}DRY-RUN: ${NORM}"
+    else
+      namespace=$(sudo docker inspect -f '{{.State.Pid}}' core$i)
+      mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' core$i)
+      echo "core$i,$namespace,$mgmt_ip" >> $container_record
+      echo -en "\t"; sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
+      a_cores["core$i"]="$namespace"
+  fi # if [ $dry_run -eq 1 ]
   done
 fi # if [ $cores -ne 0 ]; then
 
@@ -216,11 +252,15 @@ if [ $spines -ne 0 ]; then
   do
     echo "Spine number: $i"
     docker_start_spine $i
-    namespace=$(sudo docker inspect -f '{{.State.Pid}}' spine$i)
-    mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' spine$i)
-    echo -en "\t"; sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
-    echo "spine$i,$namespace,$mgmt_ip" >> $container_record
-    a_spines["spine$i"]="$namespace"
+    if [ $dry_run -eq 1 ]; then
+      echo "${REV}DRY-RUN: ${NORM}"
+    else
+      namespace=$(sudo docker inspect -f '{{.State.Pid}}' spine$i)
+      mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' spine$i)
+      echo -en "\t"; sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
+      echo "spine$i,$namespace,$mgmt_ip" >> $container_record
+      a_spines["spine$i"]="$namespace"
+    fi # if [ $dry_run -eq 1 ]
   done
 fi # if [ $spines -ne 0 ]; then
 
@@ -228,12 +268,16 @@ if [ $leaf -ne 0 ]; then
   for ((i=1; i<=$leaf; i++))
   do
     echo "Leaf number: $i"
-    docker_start_leaf $i
-    namespace=$(sudo docker inspect -f '{{.State.Pid}}' leaf$i)
-    mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' leaf$i)
-    echo -en "\t"; sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
-    echo "leaf$i,$namespace,$mgmt_ip" >> $container_record
-    a_leaves["leaf$i"]="$namespace"
+    if [ $dry_run -eq 1 ]; then
+      echo "${REV}DRY-RUN: ${NORM}"
+    else
+      docker_start_leaf $i
+      namespace=$(sudo docker inspect -f '{{.State.Pid}}' leaf$i)
+      mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' leaf$i)
+      echo -en "\t"; sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
+      echo "leaf$i,$namespace,$mgmt_ip" >> $container_record
+      a_leaves["leaf$i"]="$namespace"
+    fi # if [ $dry_run -eq 1 ]
   done
 fi # if [ $leaf -ne 0 ]; then
 
@@ -242,20 +286,19 @@ if [ $hosts -ne 0 ]; then
   do
     echo "HOST number: $i"
     docker_start_host $i
-    namespace=$(sudo docker inspect -f '{{.State.Pid}}' host$i)
-    mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' host$i)
-    sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
-    echo "host$i,$namespace,$mgmt_ip" >> $container_record
+    if [ $dry_run -eq 1 ]; then
+      echo "${REV}DRY-RUN: ${NORM}"
+    else
+      namespace=$(sudo docker inspect -f '{{.State.Pid}}' host$i)
+      mgmt_ip=$(sudo docker inspect -f '{{.NetworkSettings.Networks.bridge.IPAddress}}' host$i)
+      sudo ln -vs /proc/$namespace/ns/net /var/run/netns/$namespace
+      echo "host$i,$namespace,$mgmt_ip" >> $container_record
+    fi # if [ $dry_run -eq 1 ]
   done
 fi # if [ $leaf -ne 0 ]; then
 
 #set -x
 echo "All cores and spines:"
-#
-# The reason for different ethindex vars is that
-# on the core, the ethindex will increase unconditionally.
-# But on each spine, the index will reset.
-
 #
 # This function expects one parameter - the Docker namespace, into which it will attempt to delve
 # and suss out the next ethX number to make, and then return that number
@@ -300,7 +343,11 @@ for spine_key in "${!a_spines[@]}"; do
     core_namespace=${a_cores[$core_key]}
     echo -e "\tCore: \"$core_key\", namespace: \"$core_namespace\"";
     echo -e "\t\tCreating VETH interface between spine $spine_key and core $core_key"
-    make_veth $spine_namespace $core_namespace
+    if [ $dry_run -eq 1 ]; then
+      echo "${REV}DRY-RUN: ${NORM}"
+    else
+      make_veth $spine_namespace $core_namespace
+    fi # if [ $dry_run -eq 1 ]
 #    for key in "${!a_leaves[@]}"; do
 #      echo -e "\t\tLeaf: \"$key\", namespace: \"${a_leaves[$key]}\"";
 #    done
@@ -324,61 +371,84 @@ group2_leaves="5 6 7 8"
 #  echo $key -- ${a_spines[$key]} $group
 #done
 echo "Processing Spine/Leaf groups:"
-echo "Group 1:"
-for leaf in $group1_leaves; do
-  for spine in $group1_spines; do
-    spine_namespace=${a_spines[spine$spine]}
-    leaf_namespace=${a_leaves[leaf$leaf]}
-    #echo "Leaf ${a_leaves[leaf$leaf]} -> Spine ${a_spines[spine$spine]}"
-    echo -e "\tSpine (spine$spine, namespace:$spine_namespace) -> Leaf (leaf$leaf, namespace: $leaf_namespace)"
-    make_veth $leaf_namespace $spine_namespace
+if [ $dry_run -eq 1 ]; then
+  echo "${REV}DRY-RUN${NORM}"
+else
+  echo "Group 1:"
+  for leaf in $group1_leaves; do
+    for spine in $group1_spines; do
+      spine_namespace=${a_spines[spine$spine]}
+      leaf_namespace=${a_leaves[leaf$leaf]}
+      #echo "Leaf ${a_leaves[leaf$leaf]} -> Spine ${a_spines[spine$spine]}"
+      echo -e "\tSpine (spine$spine, namespace:$spine_namespace) -> Leaf (leaf$leaf, namespace: $leaf_namespace)"
+      make_veth $leaf_namespace $spine_namespace
+    done
   done
-done
-echo "Group 2:"
-for leaf in $group2_leaves; do
-  for spine in $group2_spines; do
-    spine_namespace=${a_spines[spine$spine]}
-    leaf_namespace=${a_leaves[leaf$leaf]}
-    #echo "Leaf ${a_leaves[leaf$leaf]} -> Spine ${a_spines[spine$spine]}"
-    echo -e "\tSpine (spine$spine, namespace:$spine_namespace) -> Leaf (leaf$leaf, namespace: $leaf_namespace)"
-    make_veth $leaf_namespace $spine_namespace
-    echo
-  done
-done
+fi # if [ $dry_run -eq 1 ]
 
-echo "Sleeping 20s to allow for docker daemons to start in the background"
-sleep 20
+echo "Group 2:"
+if [ $dry_run -eq 1 ]; then
+  echo "${REV}DRY-RUN${NORM}"
+else
+  for leaf in $group2_leaves; do
+    for spine in $group2_spines; do
+      spine_namespace=${a_spines[spine$spine]}
+      leaf_namespace=${a_leaves[leaf$leaf]}
+      #echo "Leaf ${a_leaves[leaf$leaf]} -> Spine ${a_spines[spine$spine]}"
+      echo -e "\tSpine (spine$spine, namespace:$spine_namespace) -> Leaf (leaf$leaf, namespace: $leaf_namespace)"
+      make_veth $leaf_namespace $spine_namespace
+      echo
+    done
+  done
+fi # if [ $dry_run -eq 1 ]
 
 echo -e "Start flexswtich to pick up the interfaces "
-for instance in $(sudo docker ps | awk '{print $NF}'|grep -v "NAME"); do
-  echo "##############################"
-  echo "#######\"$instance\" FS restart######"
-  echo "##############################"
-  sudo docker exec $instance sh -c "/etc/init.d/flexswitch restart" 
-  if [ $? -ne 0 ]; then
-     echo "${RED}ERROR: ${NORM}Starting a flexswitch process in docker instance \"$instance\" failed. Please check output above and fix" 1>&2
-     exit 1
-  fi # if [ $? -ne 0 ]
-  echo "Sleeping 20s to allow for the FlexSwitch daemons to start"
-  sleep 20
+if [ $dry_run -eq 1 ]; then
+  echo "${REV}DRY-RUN${NORM}"
+else
+  for instance in $(sudo docker ps | awk '{print $NF}'|grep -v "NAME"); do
+    echo "##############################"
+    echo "#######\"$instance\" FS restart######"
+    echo "##############################"
+    sudo docker exec $instance sh -c "/etc/init.d/flexswitch restart" 
+    if [ $? -ne 0 ]; then
+       echo "${RED}ERROR: ${NORM}Starting a flexswitch process in docker instance \"$instance\" failed. Please check output above and fix" 1>&2
+       exit 1
+    fi # if [ $? -ne 0 ]
+    echo "Sleeping 20s to allow for the FlexSwitch daemons to start"
+    sleep 20
 done
+fi # if [ $dry_run -eq 1 ]
 
-for instance in $(sudo docker ps | awk '{print $NF}'|grep -v "NAME"); do
+if [ $dry_run -eq 1 ]; then
   echo "Checking on the status of FlexSwitch in docker instance \"$instance\"..."
-  sudo docker exec $instance sh -c "/etc/init.d/flexswitch status"
-  if [ $? -ne 0 ]; then
-     echo "${RED}ERROR: ${NORM}Checking a flexswitch process in docker instance \"$instance\" failed. Please check output above and fix" 1>&2
-     exit 1
-  fi # if [ $? -ne 0 ]
-done
+  echo "${REV}DRY-RUN${NORM}"
+else
+  for instance in $(sudo docker ps | awk '{print $NF}'|grep -v "NAME"); do
+    echo "Checking on the status of FlexSwitch in docker instance \"$instance\"..."
+    sudo docker exec $instance sh -c "/etc/init.d/flexswitch status"
+    if [ $? -ne 0 ]; then
+       echo "${RED}ERROR: ${NORM}Checking a flexswitch process in docker instance \"$instance\" failed. Please check output above and fix" 1>&2
+       exit 1
+    fi # if [ $? -ne 0 ]
+  done
+fi # if [ $dry_run -eq 1 ]
 
 echo "Network Links:"
-IFS=","
-cat $container_record | while read cid name namespace ip; do 
-  sed -i "s/$namespace/$name/g" $netlinks 
-done
-unset IFS
-sort netlinks
+if [ $dry_run -eq 1 ]; then
+  echo "${REV}DRY-RUN${NORM}"
+else
+  IFS=","
+  cat $container_record | while read cid name namespace ip; do 
+    sed -i "s/$namespace/$name/g" $netlinks 
+  done
+  unset IFS
+  sort netlinks
+fi # if [ $dry_run -eq 1 ]
 
 echo "Containers:"
-cat $container_record
+if [ $dry_run -eq 1 ]; then
+  echo "${REV}DRY-RUN${NORM}"
+else
+  cat $container_record
+fi # if [ $dry_run -eq 1 ]
